feat: move details section to rigth
