+++ class-components/src/pages/search-page/search-page.module.css
@@ -1,0 +1,9 @@
+.container {
+  display: flex;
+}
+
+.outlet {
+  position: fixed;
+  right: 0;
+  background-color: rgb(248 134 28);
+}
